Add values for model types that have semantics
For most types, we only display the values (e.g. location) -- so we don't
need to enumerate all possible types because we don't need to query on
them.
However, in some models, specific types have well defined semantics
(e.g. contact type email is used to send notifications or contact type
direct is used for opening an on-site chat) so we need to enumerate
these types explicitly so that we can refer to them consistently.

Note that for some models, we have a mix of semantic types and
non-semantic types, e.g. the responsibility type owner grants edit
permissions but users should also be able to create new responsibilities
for the calendaring widget.

Also:
- add @classproperty decorator stolen from StackOverflow to avoid having
  to create lots of in-line namedtuples using ugly syntax

with @NerdyProjects