Allow null responsibility relation
This supports the use-case where a mappable owner wants to advertise a
responsibility (e.g. "pick up food at 7pm" or "be the musician"): we
simply create a new MappableResponsibility with a null value for the
responsible which gets displayed as "help wanted" in the UI and can
create a notification to interested users. When a user signs up for the
responsibility, the responsible field in the relation gets filled in.