Add skeleton of yunity API
We use native Django routes and views.

We also considered using DRF but it was a bit too magical: lots of
implicit stuff and conventions going, etc.

We also considered using routing decorators as an alternative using
[django-quicky](https://github.com/sametmax/django-quicky/) but the
resulting code was quite ugly because that library doesn't seem to be
built with non HTTP-GET routes in mind...

with @NerdyProjects