Add flask-swagger clone for API documentation
The flask-swagger project allows documenting a flask API.
That application is now used to parse the docstrings of all urlpatterns
and output a swagger definition file at /doc.

It should be refactored to a distinct module that works for general
django applications. Also, integrating a swagger-io instance would be
nice. For now, README describes how to integrate swagger into the django
server.