Enable cross origin requests
CORS allows urls to be requested from any other domain. This is useful
for requesting the documentation json from other servers.
This should not be of any harm because a valid session is still
necessary to request private data and an additional valid csrf token is
needed to submit any data.