Permissions lookup for listing actions on target
Implement a (for now very dumb) method to query all available actions a
user (through being a user or part of a number of groups) can have on a
thing.

For now a poc with known very bad database performance that can easily
be fixed once the model has proven and stabilized.

with @nicksellen