Use django-rest-swagger instead of drfdocs
drfdocs did not add the methods a ViewClass were generating.
django-rest-swagger works out of the box and anyway seems a bit more
like what we want. We may still want to customize it a bit later to
easily add examples and return code documentation without having yaml in
the docstring.