Don't avoid ng-click
Rationale by @nicksellen - the rationale for avoiding them was to keep all logic in js and leave templates as simple templates so logic not spread everywhere. I suggested that this is not the angular-way and it is totally fine/usual/expected to use ng-click, etc. (use react et al if you want all logic in js).