Rejig template structure, add CurrentGroup service
Added concept of layouts. In the route/state config you can now
set:

```
parent: "main"
```

And it will automatically get the topbar component added, and
put the main content in a 70% width container so it's not too wide.

This required decoupling the topbar from the page components by
adding CurrentGroup service that holds the data of the currently
selected group instead of passing in the group data via the template.