Fix store detail test to include parent route
The test needs to load all modules which define parent routes
or the route is not registered and will not be triggered.

At the moment the routes do not have a hook to check for
authenticated status, so I removed that part of the test too.

Current backend will return group info without auth, but not
store info. Not sure what is intended.