Breadcrumbs & Pagetitle (#246)
* Use more ES6-like style

* Set max-height depending on the input

* Added ncy-angular-breadcrumbs

* Added correct names for breadcrumbs

* Added Page Title

* Fixed lint errors

* Moved setPageTitle to rootScope

* Add confirm dialog before deleting pickup

Changed the icon to a trash bin.
Use pre-rendered DOM for dialog, it seems easier if we don't need
dynamic content in the dialog.

* Move md-dialog on top of leaflet

they have astronomic integers there...
Closes #226

* Extend test

* Added simicolon to in app.js

* Use dot as seperator and name Foodsaving

dot has less ink than |
I think foodsaving is the proper name of the tool.

* Move title config to separate file

Easier to keep the overview in app.js

* Use ngInject instead of explicit annotation

* Use provided directive to generate title

It seems much easier. It's not so customizable, though. So it's reverse
now.. maybe we fix this later.

* Revert "Move md-dialog on top of leaflet"

This reverts commit 97e845e186797f4c8bca6e83328ab30d04faa076.

* Move leaflet controls below md-dialog

* Extend group.join with a data parameter

* Add lock icon

* WIP: group join with password

* Ask for group password and show public_description

* Add group settings for public_description and password

* Move toggle to HTML

* Disable button if form is invalid

Also requires that the server check status is reset after keypress

* Fix generator

Authentication is needed for the redirect hook

* Mail verification page

Tests still fail, I don't know why it doesn't find the route.

* Add password reset

TODO: three route tests are skipped, can't get it running

* Password change field on user page

* Fix tests

HAH! Go away, httpBackend! It's not that I don't like you, but it's it
simply didn't work and I still have no idea why. Plain promises to the
rescue!

* Empty controller for verifyMail

* Add automatic loading bar

Should enhance UX by showing the user that something is happening in the
background

Closes #200

* Fix spelling

* chore(package): update karma-webpack to version 2.0.0 (#261)

https://greenkeeper.io/

* Update German locale from transifex (#262)

Thanks @djahnie for the Translation!

* localize: swedish from transifex

* localize: add swedish configuration

* chore: change deployment to foodsaving.world

* fix(package): update angular-material to version 1.1.3 (#269)

Closes #267

https://greenkeeper.io/

* Breadcrumbs are now pretty (: