Add ability to delete pickups
I made it so deleting series pickups is not available because
I don't quite understand what the idea for this is right now...

Actually, in general modifying series pickups is not clear to me.
It seems quite useful in many cases (e.g. you know one week that
the pickup is not available, or will be an hour later). But I did
not think very much about the other cases, e.g. when there are
already people signed up for it.

I shouldn't be writing all this in here, but I am on the beers, so
feeling chatty.

I'm wondering whether to get another beer, but will it impact my
coding ability? I'm enjoying the music and the mood and still got
58% of my battery left. It seems I added some useful things in the
last hour, so many more beer helps.

Or will I go over the beer-productivity curve and become less
productive.

I heard this is disputed though.

Lets finish up this commit though anyway.