Upgrade to Quasar 0.15 (without new CLI) (#974)
Closes #913 

I upgraded Quasar to the latest version. It needed many changes, but it works and looks almost the same as before :smile: 

I didn't upgrade to the new CLI and the new project structure yet, as this requires moving and rewriting lots of files. But we might do it in future.

I didn't set up Quasar i18n, because we don't use any component that needs it: http://quasar-framework.org/components/internationalization.html


* WIP quasar 0.15

* more quasar upgrading

* fix dialogs, use appVisibility plugin

* use notify plugin

* use native events for collapsible cards

* use v-close-overlay instead of refs

* use new dialog style

* fix pickup edit

* fix history detail toggle

* fix loadingprogress bar

* start fixing tests

* fix test interactions

* update snapshots

* fix code style

* fix design

* hide give feedback button if no feedback can be given

* fix drawer

* fix tablet layout

* update snapshots

* fix gallery style

* fix conversation compose style

* fix test

* update locales

* update snapshots