fix cordova.js injection
It seems it was missing since the webpack 4 upgrade. We don't rely on it
very much, that's why we didn't notice it in the beginning. Most
notably, the deviceready event doesn't get fired, which leads to missing
FCM.
I added a warning if
the deviceready event didn't get fired within 5 seconds, to make it
easier to detect in future.