Conversation: handle on-screen keyboard
textarea focus and keyboard open are not coupled, i.e. the textarea can have focus, but the keyboard can be closed.
We also need to make sure to not scroll away from the cursor or even hide it.
This can easily happen when the textarea becomes bigger than the screen size.
The best trick seems to call blur() and focus() and let the browser do the rest.