Added duplicate emoji fix. Changed emoji matching.
We had an issue where the same emoji with different names would
show up twice on the search results. We now check a set to avoid
duplicates. Also, we changed the emoji matching regular expression
to match substrings instead of subsequences to avoid results that
the user might not have intended to search for.