Emoji picker search bar (#1421)
* Added a search bar without functionality

This is the first step towards making the emoji picker select the emojis
the user inputs. Used QSearch from quasar.

* Added exact search functionality. Breaks when twemoji tries to parse undefined emoji name.

* Working exact search implementation. Problem with certain emoji (e.g. fr). Currently debugging.

* Added a check that returns the 'carrot' emoji if the requested emoji cannot be found in emojiList. This is to prevent an undefined Node error if the backend normalizes the emoji name to one that differs from markdown-it-emoji's list.

* Hardcoded 5 more arbitrary emoji that we believe to be useful

* Moved search bar from ConversationAddReaction to ConversationAddReactionInner

* Added partial result matching to the picker

Currently, we use regular expressions to find matching emoji.
We decided to say that an emoji matched if the string was a subsequence
of the emoji name. Also added the functionality of limiting results to 20.

* Added duplicate emoji fix. Changed emoji matching.

We had an issue where the same emoji with different names would
show up twice on the search results. We now check a set to avoid
duplicates. Also, we changed the emoji matching regular expression
to match substrings instead of subsequences to avoid results that
the user might not have intended to search for.

* Changed instances of let/var to const

* Using includes instead of match for clarity

Before, we were using regular expressions to match the
shortcodes. The includes method is simpler now. Also,
instead of using the index of the shortcodes on the loop,
we now loop over the shortcodes directly. This was done
to make the code simpler too.

* Moved inline style to style section

Doing this to minimize inline styles

* Added two tests for the search bar

The tests try to follow the same style as the others.
The first test tests that when the 'search' variable changes,
the emoji displayed also change. The second test test that
when searches yield many results, the number of emoji displayed
are truncated to 20.

* Changed the variable name whitelist to results

It is not just a whitelist anymore, so results is more appropriate.

* Replaced row with k-add-reactions

This way, we're not modifying quasar's classes

* Moved the width of the popover to the container

Now the popover never exits the screen. We're avoiding a race
condition between Qpopover and Vue rendering the inner component.

* Refreshed snapshots

* Refreshed snapshots in UTC

* Removed carrot workaround

* Correctly reverted changes for carrot workaround

Co-authored-by: Michael Zhong <zhongmic@umich.edu>