Use address autocomplete (#1878)
* Use address autocomplete

As it used to be.
Also defaults place edit map to be group centre if available.

* Remove uneeded extra wrapped div

* Update setup for storyshots testing

"sync: false" seems to be the new way it goes and will be the default
in the future, see https://github.com/vuejs/vue-test-utils/issues/1137

The transition group stubs seem to be there by default now too, I think..
Anyway it works where it did no tbefore :)

Also see https://github.com/vuejs/vue-test-utils/issues/1163

* Add @types/jest - for editor support

See https://blog.jetbrains.com/webstorm/2018/10/testing-with-jest-in-webstorm/

* Increase test wait

Not really sure why this changed, but don't want to think about
it for now...

* Add text to translation file

* Use the spelling of the enemy

* Simplify template logic