remove broken currentThread/setMuted mutation
It tries to mutate a frozen object, which was introduced when we started freezing objects in vuex for performance reasons.

Despite this, the site worked fine for the user, because websocket updates did the job. I take this chance to remove the broken code, further embracing websockets for more essential page functionality.