Improve testing (#2536)
This is a change of testing approach, loosely based on Kent C. Dodds blog post Write tests. Not too many. Mostly integration..

The premise is that the tooling is good enough to write integration tests that run reliably, and are easy enough to write. The older paradigm of a bazillion unit tests was partly based on it being difficult to write integration tests.

So, the new approach is:

    use https://testing-library.com which encourages higher level testing from the user perspective
    two types of tests:
        page-level tests (e.g. testing stuff in src/*/pages/*.vue) that load the page component but not the router
        App tests, that load the root App component and the router
    use the new mockBackend which mocks at the axios-layer, for all tests that need it


Co-authored-by: Tilmann Becker <tilmann.becker@gmail.com>